requirements.txt
colorama==0.3.7
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.7
django-toolbelt==0.0.1
djangorestframework==3.3.3
gunicorn==19.6.0
httpie==0.9.4
mysqlclient==1.3.7
psycopg2==2.6.2
py==1.4.31
Pygments==2.1.3
requests==2.10.0
static3==0.7.0
whitenoise==3.2
